requirements.txt

openai-whisper>=20231117
opencc-python-reimplemented>=1.1.0
pydub>=0.25.1
PySide6>=6.6.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
moviepy=1.0.3
PySide6-Addons